refactor(map): collapse to a single default map with layers, basemaps and tools
Replace the multi-map composition flow with a singleton default map. The
module now exposes one map that the user can configure: layers, basemaps and
tools are added directly to it. Layer source / basemap catalog management
moves to the dedicated /map/layers screen.

- Add MapConstants.DefaultMapId and lazy seed via MapModuleOptions
- IMapContracts: replace map CRUD with GetDefault/UpdateDefault
- New /api/map/default GET (Map.View) and PUT (Map.Update) endpoints
- Browse page renders the default map with editable layer/basemap panel,
  tool buttons (export PNG, geolocate) and basemap switcher
- Drop View/Edit pages and endpoints; drop CreateMapRequest/UpdateMapRequest
- Update tests to cover default map seed, upsert, validation and authz
