+++ modules/Map/src/SimpleModule.Map/Pages/index.ts
@@ -1,6 +1,4 @@
 export const pages: Record<string, unknown> = {
   'Map/Browse': () => import('./Browse'),
   'Map/Layers': () => import('./Layers'),
-  'Map/View': () => import('./View'),
-  'Map/Edit': () => import('./Edit'),
 };
